fixed minor error

+++ NodeComm/NodeComm_ReplicaLogic.go
@@ -278,7 +278,7 @@ func (n *Node) electionTimer() {
 	//Send out electSelf messages
 	for _, pRec := range n.peerRecords {
 		if pRec.Id > n.myPRecord.Id {
-			n.DispatchCoordinationMessage(pRec, &pc.CoordinationMessage{Type: pc.CoordinationMessage_ElectSelf, Spare: len(n.outstandingFiles)})
+			n.DispatchCoordinationMessage(pRec, &pc.CoordinationMessage{Type: pc.CoordinationMessage_ElectSelf, Spare: int32(len(n.outstandingFiles))})
 		}
 	}
 	//Wait for election to finish
